+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## 0.1.2
+
+Added showsBackgroundLocationIndicator to iOS fixing [#29](https://github.com/Almoullim/background_location/issues/30#issuecomment-668540916)
+Update permission handler to v5.0.1+1
+
+## 0.1.1
+
+Bug fixes
+
+## 0.1.0
+
+- `Location` class is now public and can imported
+- `Location` class now contains a `toMap()` that will return JSON compatible map
+
 ## 0.0.11
 
 Bug fixes